Added details box and preset combobox
Added a .Text for the description field, as .Entry won't allow multiple
lines.
Also added a preset combobox on the right subframe. In it, user can
choose different saving presets.
Created the default.json file, where presets are read from.
Made ui_main.py read the default.json file.
Finally, added a "show_categories" function that .destroy the labels.
Still pending on the funtion to create the new lables.
From now on, won't be able to work in the proyect for a while as I have
1 month left for my JLPT N1 test (December 7, 2025)